% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastsk_compute_kernel}
\alias{fastsk_compute_kernel}
\title{FastSK: A Fast and Accurate GKM-SVM}
\usage{
fastsk_compute_kernel(
  train_file,
  test_file,
  out_file,
  g,
  m,
  t = 1L,
  approx = FALSE,
  delta = 0.025,
  max_iters = 100L,
  skip_variance = FALSE,
  dictionary_file = ""
)
}
\arguments{
\item{train_file}{A FASTA file containing training sequences and their label}

\item{test_file}{A FASTA file containing testing sequences and their label}

\item{g}{The length of the substrings used to compare sequences. Constraints: \code{0 < g < 20}}

\item{m}{The maximum number of mismatches when comparing two gmers Constraints: \code{0 <= m < g}}

\item{t}{The number of threads to used. Default is 1}

\item{approx}{A boolean; if set to true, then the fast approximation algorithm is used to compute the kernel. Default is False}

\item{delta}{Default is 0.025}

\item{max_iters}{The maximum number of iterations to run. Default is 100}

\item{skip_variance}{Default is False}

\item{dictionary_file}{A file containing the alphabet of characters appearing in the sequences. If not provided, the dictionary will be inferred}
}
\description{
Computes the kernel for a GKM-SVM
}
