% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastsk_train_and_score}
\alias{fastsk_train_and_score}
\title{FastSK: A Fast and Accurate GKM-SVM}
\usage{
fastsk_train_and_score(
  train_file,
  test_file,
  g,
  m,
  t = 1L,
  approx = FALSE,
  delta = 0.025,
  max_iters = 100L,
  skip_variance = FALSE,
  C = 1,
  nu = 1,
  eps = 1,
  kernel_type = "linear",
  dictionary_file = "",
  metric = "auc",
  metric_file = "auc_file.txt"
)
}
\arguments{
\item{train_file}{A FASTA file containing training sequences and their label}

\item{test_file}{A FASTA file containing testing sequences and their label}

\item{g}{The length of the substrings used to compare sequences. Constraints: \code{0 < g < 20}}

\item{m}{The maximum number of mismatches when comparing two gmers Constraints: \code{0 <= m < g}}

\item{t}{The number of threads to used. Default is 1}

\item{approx}{A boolean; if set to true, then the fast approximation algorithm is used to compute the kernel. Default is False}

\item{delta}{A numerical constant for early stopping of kernel calculation. If skip_variance is False,
the kernel calculation will terminate when \code{delta / stdv > 1.96}. Default is 0.025}

\item{max_iters}{The maximum number of iterations to run. Default is 100}

\item{skip_variance}{A boolean flag; if set to true, skip kernel standard deviation calculations and run until
max_iters is reached. Default is False}

\item{C}{SVM C parameter. Default is 1.0}

\item{nu}{SVM nu parameter. Default is 1.0}

\item{eps}{SVM epsilon parameter. Default is 1.0}

\item{kernel_type}{The kernel type to used. Must be one of linear (default), fastsk, or rbf}

\item{dictionary_file}{A file containing the alphabet of characters appearing in the sequences.
If not provided, the dictionary will be inferred}

\item{metric}{The metric to score the test sequences with. Must be one of auc (default) or accuracy}

\item{metric_file}{A filepath to write the test sequence predictions. Default is auc_file.txt}
}
\description{
Trains a fast and accurate gkm-svm and scores on test sequences
}
